requirements.txt
# Reddit scraping
snscrape>=0.7.0.20230622

# HTTP requests for NewsAPI
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Google Sheets API
gspread>=5.12.0
google-auth>=2.23.0

# Timezone handling
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
